% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_cML_Overlap.R
\name{mr_cML_Overlap}
\alias{mr_cML_Overlap}
\title{MRcML method for overlapping samples}
\usage{
mr_cML_Overlap(
  b_exp,
  b_out,
  se_exp,
  se_out,
  K_vec = 0:(length(b_exp) - 2),
  random_start = 0,
  maxit = 100,
  random_seed = 0,
  n,
  rho = 0
)
}
\arguments{
\item{b_exp}{Vector of estimated effects for exposure.}

\item{b_out}{Vector or estimated effects for outcome.}

\item{se_exp}{Vector of standard errors for exposure.}

\item{se_out}{Vector of standard errors for outcome.}

\item{K_vec}{Sets of candidate K's, the constraint parameter representing number of invalid IVs.}

\item{random_start}{Number of random start points for cML, default is 0.}

\item{maxit}{Maximum number of iterations for each optimization.}

\item{random_seed}{Random seed, an integer. Default is
0, which does not set random seed; user could specify a positive integer
as random seed to get replicable results.}

\item{n}{Sample size.}

\item{rho}{Correlation between GWAS summary statistics due to overlapping samples.}
}
\value{
A list contains full results of cML methods.
MA_BIC_theta, MA_BIC_se, MA_BIC_p:
Estimate of theta,
its standard error and p-value from cML-MA-BIC.
Similarly for BIC_theta, BIC_se, BIC_p from cML-BIC.
BIC_invalid is the set of invalid IVs selected by cML-BIC,
BIC_vec is the BIC vector.
}
\description{
This is the main function of MRcML method with overlapping samples, without data perturbation.
}
